Split most functions into separate categories.
