+++ pallets/stateful-storage/src/tests/apply_item_actions_tests.rs
@@ -1,0 +1,836 @@
+use crate::{
+	stateful_child_tree::StatefulChildTree,
+	tests::{
+		mock::*,
+		test_common::{constants::*, test_constants::*},
+	},
+	types::*,
+	Config, Error, Event as StatefulEvent,
+};
+use codec::Encode;
+use common_primitives::{
+	stateful_storage::{PageHash, PageNonce},
+	utils::wrap_binary_data,
+};
+use frame_support::{assert_err, assert_ok, BoundedVec};
+#[allow(unused_imports)]
+use pretty_assertions::{assert_eq, assert_ne, assert_str_eq};
+use sp_core::{sr25519, Pair};
+use sp_runtime::MultiSignature;
+
+#[test]
+fn apply_item_actions_with_invalid_msa_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1;
+		let caller_1 = test_public(INVALID_MSA_ID); // hard-coded in mocks to return None for MSA
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions(
+				RuntimeOrigin::signed(caller_1),
+				msa_id,
+				schema_id,
+				NONEXISTENT_PAGE_HASH,
+				BoundedVec::try_from(actions).unwrap(),
+			),
+			Error::<Test>::InvalidMessageSourceAccount
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_invalid_schema_id_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1;
+		let caller_1 = test_public(msa_id);
+		let schema_id = INVALID_SCHEMA_ID;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions(
+				RuntimeOrigin::signed(caller_1),
+				msa_id,
+				schema_id,
+				NONEXISTENT_PAGE_HASH,
+				BoundedVec::try_from(actions).unwrap(),
+			),
+			Error::<Test>::InvalidSchemaId
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_invalid_schema_location_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1;
+		let caller_1 = test_public(msa_id);
+		let schema_id = PAGINATED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions(
+				RuntimeOrigin::signed(caller_1),
+				msa_id,
+				schema_id,
+				NONEXISTENT_PAGE_HASH,
+				BoundedVec::try_from(actions).unwrap(),
+			),
+			Error::<Test>::SchemaPayloadLocationMismatch
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_no_delegation_and_different_caller_from_owner_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1;
+		let caller_1 = test_public(2);
+		let schema_id = UNDELEGATED_ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions(
+				RuntimeOrigin::signed(caller_1),
+				msa_id,
+				schema_id,
+				NONEXISTENT_PAGE_HASH,
+				BoundedVec::try_from(actions).unwrap(),
+			),
+			Error::<Test>::UnauthorizedDelegate
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_corrupted_state_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1;
+		let caller_1 = test_public(msa_id);
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let page: ItemizedPage<Test> = generate_page(None, None);
+		let actions1 = vec![ItemAction::Add { data: payload.clone().try_into().unwrap() }];
+		let key = (schema_id,);
+		StatefulChildTree::<<Test as Config>::KeyHasher>::write(
+			&msa_id,
+			PALLET_STORAGE_PREFIX,
+			ITEMIZED_STORAGE_PREFIX,
+			&key,
+			&page,
+		);
+		let previous_page: ItemizedPage<Test> =
+			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
+				&msa_id,
+				PALLET_STORAGE_PREFIX,
+				ITEMIZED_STORAGE_PREFIX,
+				&key,
+			)
+			.unwrap()
+			.unwrap();
+		let previous_hash = previous_page.get_hash();
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions(
+				RuntimeOrigin::signed(caller_1),
+				msa_id,
+				schema_id,
+				previous_hash,
+				BoundedVec::try_from(actions1).unwrap(),
+			),
+			Error::<Test>::CorruptedState
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_initial_state_with_stale_hash_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1;
+		let caller_1 = test_public(msa_id);
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions1 = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions(
+				RuntimeOrigin::signed(caller_1),
+				msa_id,
+				schema_id,
+				1u32, // any non-zero value
+				BoundedVec::try_from(actions1).unwrap(),
+			),
+			Error::<Test>::StalePageState
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_existing_page_with_stale_hash_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1;
+		let caller_1 = test_public(msa_id);
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions1 = vec![ItemAction::Add { data: payload.clone().try_into().unwrap() }];
+
+		let page = ItemizedPage::<Test>::default();
+		let page_hash = page.get_hash();
+		let mut new_page =
+			ItemizedOperations::<Test>::apply_item_actions(&page, &actions1).unwrap();
+		new_page.nonce = 1;
+		let key = (schema_id,);
+		<StatefulChildTree>::write(
+			&msa_id,
+			PALLET_STORAGE_PREFIX,
+			ITEMIZED_STORAGE_PREFIX,
+			&key,
+			&new_page,
+		);
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions(
+				RuntimeOrigin::signed(caller_1),
+				msa_id,
+				schema_id,
+				page_hash,
+				BoundedVec::try_from(actions1).unwrap(),
+			),
+			Error::<Test>::StalePageState
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_initial_state_with_valid_input_should_update_storage() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1;
+		let caller_1 = test_public(msa_id);
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let prev_content_hash: PageHash = 0;
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+
+		// act
+		assert_ok!(StatefulStoragePallet::apply_item_actions(
+			RuntimeOrigin::signed(caller_1),
+			msa_id,
+			schema_id,
+			prev_content_hash,
+			BoundedVec::try_from(actions).unwrap(),
+		));
+
+		// assert
+		let updated_page: Option<ItemizedPage<Test>> =
+			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
+				&msa_id,
+				PALLET_STORAGE_PREFIX,
+				ITEMIZED_STORAGE_PREFIX,
+				&(schema_id,),
+			)
+			.unwrap();
+		assert!(updated_page.is_some());
+		let updated_page = updated_page.unwrap();
+		let curr_content_hash = updated_page.get_hash();
+		assert_eq!(updated_page.nonce, PageNonce::default() + 1);
+		System::assert_last_event(
+			StatefulEvent::ItemizedPageUpdated {
+				msa_id,
+				schema_id,
+				prev_content_hash,
+				curr_content_hash,
+			}
+			.into(),
+		);
+	});
+}
+
+#[test]
+fn apply_item_actions_existing_page_with_valid_input_should_update_storage() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1;
+		let caller_1 = test_public(msa_id);
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let nonce = 10;
+		let actions = vec![ItemAction::Add { data: payload.clone().try_into().unwrap() }];
+		let page = create_itemized_page_from::<Test>(Some(nonce), &[payload.try_into().unwrap()]);
+		let prev_content_hash = page.get_hash();
+		let key = (schema_id,);
+
+		// act
+		<StatefulChildTree>::write(
+			&msa_id,
+			PALLET_STORAGE_PREFIX,
+			ITEMIZED_STORAGE_PREFIX,
+			&key,
+			&page,
+		);
+		assert_ok!(StatefulStoragePallet::apply_item_actions(
+			RuntimeOrigin::signed(caller_1),
+			msa_id,
+			schema_id,
+			prev_content_hash,
+			BoundedVec::try_from(actions).unwrap(),
+		));
+
+		// assert
+		let updated_page: Option<ItemizedPage<Test>> =
+			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
+				&msa_id,
+				PALLET_STORAGE_PREFIX,
+				ITEMIZED_STORAGE_PREFIX,
+				&(schema_id,),
+			)
+			.unwrap();
+		assert!(updated_page.is_some());
+		let updated_page = updated_page.unwrap();
+		let curr_content_hash = updated_page.get_hash();
+		assert_eq!(nonce + 1, updated_page.nonce);
+		System::assert_last_event(
+			StatefulEvent::ItemizedPageUpdated {
+				msa_id,
+				schema_id,
+				prev_content_hash,
+				curr_content_hash,
+			}
+			.into(),
+		);
+	});
+}
+
+#[test]
+fn apply_item_actions_with_valid_input_and_empty_items_should_remove_storage() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let msa_id = 1;
+		let caller_1 = test_public(msa_id);
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions1 = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+		let actions2 = vec![ItemAction::Delete { index: 0 }];
+		let keys = (schema_id,);
+		assert_ok!(StatefulStoragePallet::apply_item_actions(
+			RuntimeOrigin::signed(caller_1.clone()),
+			msa_id,
+			schema_id,
+			NONEXISTENT_PAGE_HASH,
+			BoundedVec::try_from(actions1).unwrap(),
+		));
+
+		let items1: Option<ItemizedPage<Test>> =
+			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
+				&msa_id,
+				PALLET_STORAGE_PREFIX,
+				ITEMIZED_STORAGE_PREFIX,
+				&keys,
+			)
+			.unwrap();
+		assert!(items1.is_some());
+		let content_hash = items1.unwrap().get_hash();
+
+		// act
+		assert_ok!(StatefulStoragePallet::apply_item_actions(
+			RuntimeOrigin::signed(caller_1),
+			msa_id,
+			schema_id,
+			content_hash,
+			BoundedVec::try_from(actions2).unwrap(),
+		));
+
+		// assert
+		let items2: Option<ItemizedPage<Test>> =
+			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
+				&msa_id,
+				PALLET_STORAGE_PREFIX,
+				ITEMIZED_STORAGE_PREFIX,
+				&(schema_id,),
+			)
+			.unwrap();
+		assert!(items2.is_none());
+		System::assert_last_event(
+			StatefulEvent::ItemizedPageDeleted {
+				msa_id,
+				schema_id,
+				prev_content_hash: content_hash,
+			}
+			.into(),
+		);
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_wrong_signature_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let (msa_id, pair) = get_signature_account();
+		let delegator_key = pair.public();
+		let (signature_key, _) = sr25519::Pair::generate();
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions).unwrap(),
+			target_hash: PageHash::default(),
+			msa_id,
+			expiration: 10,
+			schema_id,
+		};
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = signature_key.sign(&encode_data_new_key_data).into();
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions_with_signature(
+				RuntimeOrigin::signed(caller_1),
+				delegator_key.into(),
+				owner_signature,
+				payload
+			),
+			Error::<Test>::InvalidSignature
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_too_far_expiration_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let (msa_id, pair) = get_signature_account();
+		let delegator_key = pair.public();
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions).unwrap(),
+			target_hash: PageHash::default(),
+			msa_id,
+			expiration: (<Test as Config>::MortalityWindowSize::get() + 1).into(),
+			schema_id,
+		};
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions_with_signature(
+				RuntimeOrigin::signed(caller_1),
+				delegator_key.into(),
+				owner_signature,
+				payload
+			),
+			Error::<Test>::ProofNotYetValid
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_expired_payload_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let (msa_id, pair) = get_signature_account();
+		let delegator_key = pair.public();
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+		let block_number = 10;
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions).unwrap(),
+			target_hash: PageHash::default(),
+			msa_id,
+			expiration: block_number,
+			schema_id,
+		};
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
+
+		// act
+		System::set_block_number(block_number);
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions_with_signature(
+				RuntimeOrigin::signed(caller_1),
+				delegator_key.into(),
+				owner_signature,
+				payload
+			),
+			Error::<Test>::ProofHasExpired
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_correct_input_should_work() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let (msa_id, pair) = get_signature_account();
+		let delegator_key = pair.public();
+		let schema_id = ITEMIZED_SCHEMA;
+		let prev_content_hash = PageHash::default();
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions).unwrap(),
+			target_hash: prev_content_hash,
+			msa_id,
+			expiration: 10,
+			schema_id,
+		};
+
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
+
+		// act
+		assert_ok!(StatefulStoragePallet::apply_item_actions_with_signature(
+			RuntimeOrigin::signed(caller_1),
+			delegator_key.into(),
+			owner_signature,
+			payload
+		));
+
+		// assert
+		let updated_page: Option<ItemizedPage<Test>> =
+			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
+				&msa_id,
+				PALLET_STORAGE_PREFIX,
+				ITEMIZED_STORAGE_PREFIX,
+				&(schema_id,),
+			)
+			.unwrap();
+		assert!(updated_page.is_some());
+		let curr_content_hash = updated_page.unwrap().get_hash();
+		System::assert_last_event(
+			StatefulEvent::ItemizedPageUpdated {
+				msa_id,
+				schema_id,
+				prev_content_hash,
+				curr_content_hash,
+			}
+			.into(),
+		);
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_non_existing_msa_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let pair = get_invalid_msa_signature_account(); // hardcoded key that returns None Msa
+		let delegator_key = pair.public();
+		let msa_id = 1;
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions).unwrap(),
+			target_hash: PageHash::default(),
+			msa_id,
+			expiration: 10,
+			schema_id,
+		};
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions_with_signature(
+				RuntimeOrigin::signed(caller_1),
+				delegator_key.into(),
+				owner_signature,
+				payload
+			),
+			Error::<Test>::InvalidMessageSourceAccount
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_wrong_msa_in_payload_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let (_correct_msa_id, pair) = get_signature_account();
+		let delegator_key = pair.public();
+		let wrong_msa_id = 3;
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions).unwrap(),
+			target_hash: PageHash::default(),
+			msa_id: wrong_msa_id,
+			expiration: 10,
+			schema_id,
+		};
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions_with_signature(
+				RuntimeOrigin::signed(caller_1),
+				delegator_key.into(),
+				owner_signature,
+				payload
+			),
+			Error::<Test>::InvalidMessageSourceAccount
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_invalid_schema_id_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let (msa_id, pair) = get_signature_account();
+		let delegator_key = pair.public();
+		let schema_id = INVALID_SCHEMA_ID;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions).unwrap(),
+			target_hash: PageHash::default(),
+			msa_id,
+			expiration: 10,
+			schema_id,
+		};
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions_with_signature(
+				RuntimeOrigin::signed(caller_1),
+				delegator_key.into(),
+				owner_signature,
+				payload
+			),
+			Error::<Test>::InvalidSchemaId
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_invalid_schema_location_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let (msa_id, pair) = get_signature_account();
+		let delegator_key = pair.public();
+		let schema_id = PAGINATED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions).unwrap(),
+			target_hash: PageHash::default(),
+			msa_id,
+			expiration: 10,
+			schema_id,
+		};
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions_with_signature(
+				RuntimeOrigin::signed(caller_1),
+				delegator_key.into(),
+				owner_signature,
+				payload
+			),
+			Error::<Test>::SchemaPayloadLocationMismatch
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_page_with_stale_hash_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let (msa_id, pair) = get_signature_account();
+		let delegator_key = pair.public();
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions = vec![ItemAction::Add { data: payload.clone().try_into().unwrap() }];
+		let page = ItemizedPage::<Test>::default();
+		let page_hash = page.get_hash();
+		let page = ItemizedOperations::<Test>::apply_item_actions(&page, &actions).unwrap();
+		let key = (schema_id,);
+		<StatefulChildTree>::write(
+			&msa_id,
+			PALLET_STORAGE_PREFIX,
+			ITEMIZED_STORAGE_PREFIX,
+			&key,
+			page,
+		);
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions).unwrap(),
+			target_hash: page_hash,
+			msa_id,
+			expiration: 10,
+			schema_id,
+		};
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions_with_signature(
+				RuntimeOrigin::signed(caller_1),
+				delegator_key.into(),
+				owner_signature,
+				payload
+			),
+			Error::<Test>::StalePageState
+		)
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_valid_input_and_empty_items_should_remove_storage() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let (msa_id, pair) = get_signature_account();
+		let delegator_key = pair.public();
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let actions1 = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
+		let actions2 = vec![ItemAction::Delete { index: 0 }];
+		let keys = (schema_id,);
+		assert_ok!(StatefulStoragePallet::apply_item_actions(
+			RuntimeOrigin::signed(caller_1.clone()),
+			msa_id,
+			schema_id,
+			NONEXISTENT_PAGE_HASH,
+			BoundedVec::try_from(actions1).unwrap(),
+		));
+
+		let items1: Option<ItemizedPage<Test>> =
+			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
+				&msa_id,
+				PALLET_STORAGE_PREFIX,
+				ITEMIZED_STORAGE_PREFIX,
+				&keys,
+			)
+			.unwrap();
+		assert!(items1.is_some());
+		let content_hash = items1.unwrap().get_hash();
+
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions2).unwrap(),
+			target_hash: content_hash,
+			msa_id,
+			expiration: 10,
+			schema_id,
+		};
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
+
+		// act
+		assert_ok!(StatefulStoragePallet::apply_item_actions_with_signature(
+			RuntimeOrigin::signed(caller_1),
+			delegator_key.into(),
+			owner_signature,
+			payload
+		));
+
+		// assert
+		let items2: Option<ItemizedPage<Test>> =
+			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
+				&msa_id,
+				PALLET_STORAGE_PREFIX,
+				ITEMIZED_STORAGE_PREFIX,
+				&(schema_id,),
+			)
+			.unwrap();
+		assert!(items2.is_none());
+		System::assert_last_event(
+			StatefulEvent::ItemizedPageDeleted {
+				msa_id,
+				schema_id,
+				prev_content_hash: content_hash,
+			}
+			.into(),
+		);
+	});
+}
+
+#[test]
+fn apply_item_actions_with_signature_having_corrupted_state_should_fail() {
+	new_test_ext().execute_with(|| {
+		// arrange
+		let caller_1 = test_public(1);
+		let (msa_id, pair) = get_signature_account();
+		let delegator_key = pair.public();
+		let schema_id = ITEMIZED_SCHEMA;
+		let payload = vec![1; 5];
+		let page: ItemizedPage<Test> = generate_page(None, None);
+		let actions = vec![ItemAction::Add { data: payload.clone().try_into().unwrap() }];
+		let key = (schema_id,);
+		StatefulChildTree::<<Test as Config>::KeyHasher>::write(
+			&msa_id,
+			PALLET_STORAGE_PREFIX,
+			ITEMIZED_STORAGE_PREFIX,
+			&key,
+			&page,
+		);
+		let previous_page: ItemizedPage<Test> =
+			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
+				&msa_id,
+				PALLET_STORAGE_PREFIX,
+				ITEMIZED_STORAGE_PREFIX,
+				&key,
+			)
+			.unwrap()
+			.unwrap();
+		let previous_hash = previous_page.get_hash();
+
+		let payload = ItemizedSignaturePayload {
+			actions: BoundedVec::try_from(actions).unwrap(),
+			target_hash: previous_hash,
+			msa_id,
+			expiration: 10,
+			schema_id,
+		};
+		let encode_data_new_key_data = wrap_binary_data(payload.encode());
+		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
+
+		// act
+		assert_err!(
+			StatefulStoragePallet::apply_item_actions_with_signature(
+				RuntimeOrigin::signed(caller_1),
+				delegator_key.into(),
+				owner_signature,
+				payload
+			),
+			Error::<Test>::CorruptedState
+		)
+	});
+}
