+++ pallets/stateful-storage/src/tests/mod.rs
@@ -2,4 +2,7 @@ pub mod mock;
 #[cfg(any(feature = "runtime-benchmarks", test))]
 mod test_common;
 
+mod apply_item_actions_tests;
+mod delete_page_tests;
 mod other_tests;
+mod upsert_page_tests;
