+++ pallets/stateful-storage/src/tests/other_tests.rs
@@ -1,544 +1,24 @@
 use crate::{
-	pallet,
 	stateful_child_tree::StatefulChildTree,
-	tests::{mock::*, test_common::constants::*},
+	tests::{
+		mock::*,
+		test_common::{constants::*, test_constants::*},
+	},
 	types::*,
-	Config, Error, Event as StatefulEvent,
+	Config, Error,
 };
-use codec::{Decode, Encode, MaxEncodedLen};
+use codec::{Encode, MaxEncodedLen};
 use common_primitives::{
 	schema::{ModelType, PayloadLocation, SchemaId},
-	stateful_storage::{PageHash, PageId, PageNonce},
+	stateful_storage::PageNonce,
 	utils::wrap_binary_data,
 };
-use frame_support::{assert_err, assert_ok, assert_storage_noop, traits::Len, BoundedVec};
+use frame_support::{assert_err, assert_ok, traits::Len, BoundedVec};
 #[allow(unused_imports)]
 use pretty_assertions::{assert_eq, assert_ne, assert_str_eq};
-use scale_info::TypeInfo;
-use sp_core::{sr25519, ConstU32, Get, Pair};
+use sp_core::{ConstU32, Pair};
 use sp_runtime::MultiSignature;
-use sp_std::{collections::btree_set::BTreeSet, hash::Hasher};
-use twox_hash::XxHash64;
-
-type ItemizedPageSize = <Test as Config>::MaxItemizedPageSizeBytes;
-type PaginatedPageSize = <Test as Config>::MaxPaginatedPageSizeBytes;
-type ItemizedBlobSize = <Test as Config>::MaxItemizedBlobSizeBytes;
-
-const NONEXISTENT_PAGE_HASH: u32 = 0;
-
-fn generate_payload_bytes<T: Get<u32>>(id: Option<u8>) -> BoundedVec<u8, T> {
-	let value = id.unwrap_or(1);
-	format!("{{'type':{value}, 'description':'another test description {value}'}}")
-		.as_bytes()
-		.to_vec()
-		.try_into()
-		.unwrap()
-}
-
-fn generate_page<T: Get<u32>>(in_nonce: Option<PageNonce>, id: Option<u8>) -> Page<T> {
-	let nonce = in_nonce.unwrap_or_default();
-	Page::<T> { nonce, data: generate_payload_bytes(id) }
-}
-
-fn add_itemized_payload_to_buffer<T: Config>(buffer: &mut Vec<u8>, payload: &[u8]) {
-	buffer.extend_from_slice(&ItemHeader { payload_len: payload.len() as u16 }.encode()[..]);
-	buffer.extend_from_slice(payload);
-}
-
-fn create_itemized_page_from<T: pallet::Config>(
-	nonce_in: Option<PageNonce>,
-	payloads: &[BoundedVec<u8, ItemizedBlobSize>],
-) -> ItemizedPage<T> {
-	let nonce = nonce_in.unwrap_or_default();
-	let mut buffer: Vec<u8> = vec![];
-	for p in payloads {
-		add_itemized_payload_to_buffer::<T>(&mut buffer, p.as_slice());
-	}
-	let data = BoundedVec::<u8, T::MaxItemizedPageSizeBytes>::try_from(buffer).unwrap();
-	ItemizedPage::<T> { nonce, data }
-}
-
-fn hash_payload(data: &[u8]) -> PageHash {
-	let mut hasher = XxHash64::with_seed(0);
-	sp_std::hash::Hash::hash(&data, &mut hasher);
-	let value_bytes: [u8; 4] =
-		hasher.finish().to_be_bytes()[..4].try_into().expect("incorrect hash size");
-	PageHash::from_be_bytes(value_bytes)
-}
-
-#[derive(Clone, Encode, Decode, PartialEq, Debug, TypeInfo, MaxEncodedLen)]
-/// A structure defining a Schema
-struct TestStruct {
-	pub model_type: ModelType,
-	pub payload_location: PayloadLocation,
-	pub number: u64,
-}
-
-#[test]
-fn upsert_page_id_out_of_bounds_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = 1;
-		let page_id = <Test as Config>::MaxPaginatedPageId::get() + 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-
-		assert_err!(
-			StatefulStoragePallet::upsert_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id.into(),
-				schema_id,
-				page_id,
-				hash_payload(&payload),
-				payload
-			),
-			Error::<Test>::PageIdExceedsMaxAllowed
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_invalid_msa_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = INVALID_MSA_ID;
-		let caller_1 = test_public(msa_id); // hard-coded in mocks to return None for MSA
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-
-		assert_err!(
-			StatefulStoragePallet::upsert_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id.into(),
-				schema_id,
-				page_id,
-				hash_payload(&payload),
-				payload
-			),
-			Error::<Test>::InvalidMessageSourceAccount
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_invalid_schema_id_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = INVALID_SCHEMA_ID;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-
-		assert_err!(
-			StatefulStoragePallet::upsert_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id.into(),
-				schema_id,
-				page_id,
-				hash_payload(&payload),
-				payload
-			),
-			Error::<Test>::InvalidSchemaId
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_invalid_schema_payload_location_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = ITEMIZED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-
-		assert_err!(
-			StatefulStoragePallet::upsert_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id.into(),
-				schema_id,
-				page_id,
-				hash_payload(&payload),
-				payload
-			),
-			Error::<Test>::SchemaPayloadLocationMismatch
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_no_delegation_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let caller_1 = test_public(1);
-		let msa_id = 2;
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-
-		assert_err!(
-			StatefulStoragePallet::upsert_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				page_id,
-				PageHash::default(),
-				payload
-			),
-			Error::<Test>::UnauthorizedDelegate
-		)
-	})
-}
-
-#[test]
-fn upsert_new_page_with_bad_state_hash_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-
-		assert_err!(
-			StatefulStoragePallet::upsert_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id.into(),
-				schema_id,
-				page_id,
-				hash_payload(&payload),
-				payload
-			),
-			Error::<Test>::StalePageState
-		)
-	})
-}
-
-#[test]
-fn upsert_existing_page_with_bad_state_hash_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 1;
-		let page = generate_page::<PaginatedPageSize>(None, Some(100));
-
-		let key = (schema_id, page_id);
-		<StatefulChildTree>::write(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			PAGINATED_STORAGE_PREFIX,
-			&key,
-			&page,
-		);
-
-		assert_err!(
-			StatefulStoragePallet::upsert_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				page_id,
-				0u32,
-				page.data,
-			),
-			Error::<Test>::StalePageState
-		)
-	})
-}
-
-#[test]
-fn upsert_new_page_succeeds() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let page: PaginatedPage<Test> = payload.clone().into();
-
-		assert_ok!(StatefulStoragePallet::upsert_page(
-			RuntimeOrigin::signed(caller_1),
-			msa_id,
-			schema_id,
-			page_id,
-			0u32,
-			payload.into()
-		));
-
-		let keys = (schema_id, page_id);
-		let new_page: PaginatedPage<Test> = <StatefulChildTree>::try_read(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			PAGINATED_STORAGE_PREFIX,
-			&keys,
-		)
-		.unwrap()
-		.expect("new page is empty");
-		assert_eq!(page.data, new_page.data, "new page contents incorrect");
-		assert_eq!(new_page.nonce, 1, "new page nonce incorrect");
-	})
-}
-
-#[test]
-fn upsert_existing_page_modifies_page() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id: SchemaId = PAGINATED_SCHEMA;
-		let page_id: PageId = 1;
-		let old_nonce = 3;
-		let old_content = generate_payload_bytes(Some(200));
-		let new_content = generate_payload_bytes::<PaginatedPageSize>(Some(201));
-		let mut old_page: PaginatedPage<Test> = old_content.clone().into();
-		old_page.nonce = old_nonce;
-
-		let keys = (schema_id, page_id);
-		<StatefulChildTree>::write(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			PAGINATED_STORAGE_PREFIX,
-			&keys,
-			old_page,
-		);
-		let old_page: PaginatedPage<Test> = <StatefulChildTree>::try_read(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			PAGINATED_STORAGE_PREFIX,
-			&keys,
-		)
-		.unwrap()
-		.unwrap();
-
-		assert_eq!(old_content, old_page.data);
-		assert_eq!(old_nonce, old_page.nonce);
-		assert_ok!(StatefulStoragePallet::upsert_page(
-			RuntimeOrigin::signed(caller_1),
-			msa_id,
-			schema_id,
-			page_id,
-			old_page.get_hash(),
-			new_content.clone().into()
-		));
-
-		let new_page: PaginatedPage<Test> = <StatefulChildTree>::try_read(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			PAGINATED_STORAGE_PREFIX,
-			&keys,
-		)
-		.unwrap()
-		.unwrap();
-		assert_eq!(new_content, new_page.data);
-		assert_eq!(old_nonce + 1, new_page.nonce);
-	})
-}
-
-#[test]
-fn delete_page_id_out_of_bounds_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = <Test as Config>::MaxPaginatedPageId::get() + 1;
-
-		assert_err!(
-			StatefulStoragePallet::delete_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				page_id,
-				NONEXISTENT_PAGE_HASH,
-			),
-			Error::<Test>::PageIdExceedsMaxAllowed
-		);
-	})
-}
-
-#[test]
-fn delete_page_with_invalid_msa_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let caller_1 = test_public(INVALID_MSA_ID); // hard-coded in mocks to return None for MSA
-		let msa_id = 1;
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 1;
-
-		assert_err!(
-			StatefulStoragePallet::delete_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				page_id,
-				NONEXISTENT_PAGE_HASH,
-			),
-			Error::<Test>::InvalidMessageSourceAccount
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_invalid_schema_id_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(1);
-		let schema_id = INVALID_SCHEMA_ID;
-		let page_id = 1;
-
-		assert_err!(
-			StatefulStoragePallet::delete_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				page_id,
-				NONEXISTENT_PAGE_HASH,
-			),
-			Error::<Test>::InvalidSchemaId
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_invalid_schema_payload_location_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = ITEMIZED_SCHEMA;
-		let page_id = 1;
-
-		assert_err!(
-			StatefulStoragePallet::delete_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				page_id,
-				NONEXISTENT_PAGE_HASH,
-			),
-			Error::<Test>::SchemaPayloadLocationMismatch
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_no_delegation_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(2);
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-
-		assert_err!(
-			StatefulStoragePallet::delete_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				page_id,
-				NONEXISTENT_PAGE_HASH,
-			),
-			Error::<Test>::UnauthorizedDelegate
-		)
-	})
-}
-
-#[test]
-fn delete_nonexistent_page_succeeds_noop() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 10;
-
-		assert_ok!(StatefulStoragePallet::delete_page(
-			RuntimeOrigin::signed(caller_1),
-			msa_id,
-			schema_id,
-			page_id,
-			NONEXISTENT_PAGE_HASH,
-		));
-	})
-}
-
-#[test]
-fn delete_existing_page_with_bad_hash_errors() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 11;
-		let page: PaginatedPage<Test> = generate_page(None, None);
-
-		let keys = (schema_id, page_id);
-		<StatefulChildTree>::write(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			PAGINATED_STORAGE_PREFIX,
-			&keys,
-			&page,
-		);
-
-		assert_err!(
-			StatefulStoragePallet::delete_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				page_id,
-				0u32,
-			),
-			Error::<Test>::StalePageState
-		);
-	})
-}
-
-#[test]
-fn delete_existing_page_succeeds() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 11;
-		let page: PaginatedPage<Test> = generate_page(None, None);
-		let page_hash = page.get_hash();
-		let keys = (schema_id, page_id);
-
-		<StatefulChildTree>::write(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			PAGINATED_STORAGE_PREFIX,
-			&keys,
-			&page,
-		);
-
-		assert_ok!(StatefulStoragePallet::delete_page(
-			RuntimeOrigin::signed(caller_1),
-			msa_id,
-			schema_id,
-			page_id,
-			page_hash
-		));
-
-		let page: Option<PaginatedPage<Test>> = <StatefulChildTree>::try_read(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			PAGINATED_STORAGE_PREFIX,
-			&keys,
-		)
-		.unwrap();
-		assert_eq!(page, None);
-	})
-}
+use sp_std::collections::btree_set::BTreeSet;
 
 #[test]
 fn parsing_a_well_formed_item_page_should_work() {
@@ -807,1582 +287,6 @@ fn child_tree_iterator() {
 }
 
 #[test]
-fn apply_item_actions_with_invalid_msa_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let msa_id = 1;
-		let caller_1 = test_public(INVALID_MSA_ID); // hard-coded in mocks to return None for MSA
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				NONEXISTENT_PAGE_HASH,
-				BoundedVec::try_from(actions).unwrap(),
-			),
-			Error::<Test>::InvalidMessageSourceAccount
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_invalid_schema_id_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = INVALID_SCHEMA_ID;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				NONEXISTENT_PAGE_HASH,
-				BoundedVec::try_from(actions).unwrap(),
-			),
-			Error::<Test>::InvalidSchemaId
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_invalid_schema_location_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = PAGINATED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				NONEXISTENT_PAGE_HASH,
-				BoundedVec::try_from(actions).unwrap(),
-			),
-			Error::<Test>::SchemaPayloadLocationMismatch
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_no_delegation_and_different_caller_from_owner_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let msa_id = 1;
-		let caller_1 = test_public(2);
-		let schema_id = UNDELEGATED_ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				NONEXISTENT_PAGE_HASH,
-				BoundedVec::try_from(actions).unwrap(),
-			),
-			Error::<Test>::UnauthorizedDelegate
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_corrupted_state_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let page: ItemizedPage<Test> = generate_page(None, None);
-		let actions1 = vec![ItemAction::Add { data: payload.clone().try_into().unwrap() }];
-		let key = (schema_id,);
-		StatefulChildTree::<<Test as Config>::KeyHasher>::write(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			ITEMIZED_STORAGE_PREFIX,
-			&key,
-			&page,
-		);
-		let previous_page: ItemizedPage<Test> =
-			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
-				&msa_id,
-				PALLET_STORAGE_PREFIX,
-				ITEMIZED_STORAGE_PREFIX,
-				&key,
-			)
-			.unwrap()
-			.unwrap();
-		let previous_hash = previous_page.get_hash();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				previous_hash,
-				BoundedVec::try_from(actions1).unwrap(),
-			),
-			Error::<Test>::CorruptedState
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_initial_state_with_stale_hash_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions1 = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				1u32, // any non-zero value
-				BoundedVec::try_from(actions1).unwrap(),
-			),
-			Error::<Test>::StalePageState
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_existing_page_with_stale_hash_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions1 = vec![ItemAction::Add { data: payload.clone().try_into().unwrap() }];
-
-		let page = ItemizedPage::<Test>::default();
-		let page_hash = page.get_hash();
-		let mut new_page =
-			ItemizedOperations::<Test>::apply_item_actions(&page, &actions1).unwrap();
-		new_page.nonce = 1;
-		let key = (schema_id,);
-		<StatefulChildTree>::write(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			ITEMIZED_STORAGE_PREFIX,
-			&key,
-			&new_page,
-		);
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				page_hash,
-				BoundedVec::try_from(actions1).unwrap(),
-			),
-			Error::<Test>::StalePageState
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_initial_state_with_valid_input_should_update_storage() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let prev_content_hash: PageHash = 0;
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-
-		// act
-		assert_ok!(StatefulStoragePallet::apply_item_actions(
-			RuntimeOrigin::signed(caller_1),
-			msa_id,
-			schema_id,
-			prev_content_hash,
-			BoundedVec::try_from(actions).unwrap(),
-		));
-
-		// assert
-		let updated_page: Option<ItemizedPage<Test>> =
-			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
-				&msa_id,
-				PALLET_STORAGE_PREFIX,
-				ITEMIZED_STORAGE_PREFIX,
-				&(schema_id,),
-			)
-			.unwrap();
-		assert!(updated_page.is_some());
-		let updated_page = updated_page.unwrap();
-		let curr_content_hash = updated_page.get_hash();
-		assert_eq!(updated_page.nonce, PageNonce::default() + 1);
-		System::assert_last_event(
-			StatefulEvent::ItemizedPageUpdated {
-				msa_id,
-				schema_id,
-				prev_content_hash,
-				curr_content_hash,
-			}
-			.into(),
-		);
-	});
-}
-
-#[test]
-fn apply_item_actions_existing_page_with_valid_input_should_update_storage() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let nonce = 10;
-		let actions = vec![ItemAction::Add { data: payload.clone().try_into().unwrap() }];
-		let page = create_itemized_page_from::<Test>(Some(nonce), &[payload.try_into().unwrap()]);
-		let prev_content_hash = page.get_hash();
-		let key = (schema_id,);
-
-		// act
-		<StatefulChildTree>::write(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			ITEMIZED_STORAGE_PREFIX,
-			&key,
-			&page,
-		);
-		assert_ok!(StatefulStoragePallet::apply_item_actions(
-			RuntimeOrigin::signed(caller_1),
-			msa_id,
-			schema_id,
-			prev_content_hash,
-			BoundedVec::try_from(actions).unwrap(),
-		));
-
-		// assert
-		let updated_page: Option<ItemizedPage<Test>> =
-			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
-				&msa_id,
-				PALLET_STORAGE_PREFIX,
-				ITEMIZED_STORAGE_PREFIX,
-				&(schema_id,),
-			)
-			.unwrap();
-		assert!(updated_page.is_some());
-		let updated_page = updated_page.unwrap();
-		let curr_content_hash = updated_page.get_hash();
-		assert_eq!(nonce + 1, updated_page.nonce);
-		System::assert_last_event(
-			StatefulEvent::ItemizedPageUpdated {
-				msa_id,
-				schema_id,
-				prev_content_hash,
-				curr_content_hash,
-			}
-			.into(),
-		);
-	});
-}
-
-#[test]
-fn apply_item_actions_with_valid_input_and_empty_items_should_remove_storage() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let msa_id = 1;
-		let caller_1 = test_public(msa_id);
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions1 = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-		let actions2 = vec![ItemAction::Delete { index: 0 }];
-		let keys = (schema_id,);
-		assert_ok!(StatefulStoragePallet::apply_item_actions(
-			RuntimeOrigin::signed(caller_1.clone()),
-			msa_id,
-			schema_id,
-			NONEXISTENT_PAGE_HASH,
-			BoundedVec::try_from(actions1).unwrap(),
-		));
-
-		let items1: Option<ItemizedPage<Test>> =
-			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
-				&msa_id,
-				PALLET_STORAGE_PREFIX,
-				ITEMIZED_STORAGE_PREFIX,
-				&keys,
-			)
-			.unwrap();
-		assert!(items1.is_some());
-		let content_hash = items1.unwrap().get_hash();
-
-		// act
-		assert_ok!(StatefulStoragePallet::apply_item_actions(
-			RuntimeOrigin::signed(caller_1),
-			msa_id,
-			schema_id,
-			content_hash,
-			BoundedVec::try_from(actions2).unwrap(),
-		));
-
-		// assert
-		let items2: Option<ItemizedPage<Test>> =
-			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
-				&msa_id,
-				PALLET_STORAGE_PREFIX,
-				ITEMIZED_STORAGE_PREFIX,
-				&(schema_id,),
-			)
-			.unwrap();
-		assert!(items2.is_none());
-		System::assert_last_event(
-			StatefulEvent::ItemizedPageDeleted {
-				msa_id,
-				schema_id,
-				prev_content_hash: content_hash,
-			}
-			.into(),
-		);
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_wrong_signature_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let (signature_key, _) = sr25519::Pair::generate();
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions).unwrap(),
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = signature_key.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidSignature
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_too_far_expiration_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions).unwrap(),
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: (<Test as Config>::MortalityWindowSize::get() + 1).into(),
-			schema_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::ProofNotYetValid
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_expired_payload_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-		let block_number = 10;
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions).unwrap(),
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: block_number,
-			schema_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		System::set_block_number(block_number);
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::ProofHasExpired
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_correct_input_should_work() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = ITEMIZED_SCHEMA;
-		let prev_content_hash = PageHash::default();
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions).unwrap(),
-			target_hash: prev_content_hash,
-			msa_id,
-			expiration: 10,
-			schema_id,
-		};
-
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_ok!(StatefulStoragePallet::apply_item_actions_with_signature(
-			RuntimeOrigin::signed(caller_1),
-			delegator_key.into(),
-			owner_signature,
-			payload
-		));
-
-		// assert
-		let updated_page: Option<ItemizedPage<Test>> =
-			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
-				&msa_id,
-				PALLET_STORAGE_PREFIX,
-				ITEMIZED_STORAGE_PREFIX,
-				&(schema_id,),
-			)
-			.unwrap();
-		assert!(updated_page.is_some());
-		let curr_content_hash = updated_page.unwrap().get_hash();
-		System::assert_last_event(
-			StatefulEvent::ItemizedPageUpdated {
-				msa_id,
-				schema_id,
-				prev_content_hash,
-				curr_content_hash,
-			}
-			.into(),
-		);
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_non_existing_msa_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let pair = get_invalid_msa_signature_account(); // hardcoded key that returns None Msa
-		let delegator_key = pair.public();
-		let msa_id = 1;
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions).unwrap(),
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidMessageSourceAccount
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_wrong_msa_in_payload_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (_correct_msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let wrong_msa_id = 3;
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions).unwrap(),
-			target_hash: PageHash::default(),
-			msa_id: wrong_msa_id,
-			expiration: 10,
-			schema_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidMessageSourceAccount
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_invalid_schema_id_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = INVALID_SCHEMA_ID;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions).unwrap(),
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidSchemaId
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_invalid_schema_location_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = PAGINATED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions).unwrap(),
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::SchemaPayloadLocationMismatch
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_page_with_stale_hash_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions = vec![ItemAction::Add { data: payload.clone().try_into().unwrap() }];
-		let page = ItemizedPage::<Test>::default();
-		let page_hash = page.get_hash();
-		let page = ItemizedOperations::<Test>::apply_item_actions(&page, &actions).unwrap();
-		let key = (schema_id,);
-		<StatefulChildTree>::write(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			ITEMIZED_STORAGE_PREFIX,
-			&key,
-			page,
-		);
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions).unwrap(),
-			target_hash: page_hash,
-			msa_id,
-			expiration: 10,
-			schema_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::StalePageState
-		)
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_valid_input_and_empty_items_should_remove_storage() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let actions1 = vec![ItemAction::Add { data: payload.try_into().unwrap() }];
-		let actions2 = vec![ItemAction::Delete { index: 0 }];
-		let keys = (schema_id,);
-		assert_ok!(StatefulStoragePallet::apply_item_actions(
-			RuntimeOrigin::signed(caller_1.clone()),
-			msa_id,
-			schema_id,
-			NONEXISTENT_PAGE_HASH,
-			BoundedVec::try_from(actions1).unwrap(),
-		));
-
-		let items1: Option<ItemizedPage<Test>> =
-			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
-				&msa_id,
-				PALLET_STORAGE_PREFIX,
-				ITEMIZED_STORAGE_PREFIX,
-				&keys,
-			)
-			.unwrap();
-		assert!(items1.is_some());
-		let content_hash = items1.unwrap().get_hash();
-
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions2).unwrap(),
-			target_hash: content_hash,
-			msa_id,
-			expiration: 10,
-			schema_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_ok!(StatefulStoragePallet::apply_item_actions_with_signature(
-			RuntimeOrigin::signed(caller_1),
-			delegator_key.into(),
-			owner_signature,
-			payload
-		));
-
-		// assert
-		let items2: Option<ItemizedPage<Test>> =
-			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
-				&msa_id,
-				PALLET_STORAGE_PREFIX,
-				ITEMIZED_STORAGE_PREFIX,
-				&(schema_id,),
-			)
-			.unwrap();
-		assert!(items2.is_none());
-		System::assert_last_event(
-			StatefulEvent::ItemizedPageDeleted {
-				msa_id,
-				schema_id,
-				prev_content_hash: content_hash,
-			}
-			.into(),
-		);
-	});
-}
-
-#[test]
-fn apply_item_actions_with_signature_having_corrupted_state_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = ITEMIZED_SCHEMA;
-		let payload = vec![1; 5];
-		let page: ItemizedPage<Test> = generate_page(None, None);
-		let actions = vec![ItemAction::Add { data: payload.clone().try_into().unwrap() }];
-		let key = (schema_id,);
-		StatefulChildTree::<<Test as Config>::KeyHasher>::write(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			ITEMIZED_STORAGE_PREFIX,
-			&key,
-			&page,
-		);
-		let previous_page: ItemizedPage<Test> =
-			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
-				&msa_id,
-				PALLET_STORAGE_PREFIX,
-				ITEMIZED_STORAGE_PREFIX,
-				&key,
-			)
-			.unwrap()
-			.unwrap();
-		let previous_hash = previous_page.get_hash();
-
-		let payload = ItemizedSignaturePayload {
-			actions: BoundedVec::try_from(actions).unwrap(),
-			target_hash: previous_hash,
-			msa_id,
-			expiration: 10,
-			schema_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::apply_item_actions_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::CorruptedState
-		)
-	});
-}
-
-#[test]
-fn upsert_page_with_signature_having_page_id_out_of_bounds_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = (<Test as Config>::MaxPaginatedPageId::get() + 1).into();
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let payload = PaginatedUpsertSignaturePayload {
-			payload,
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::upsert_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::PageIdExceedsMaxAllowed
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_signature_having_expired_payload_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let block_number = 10;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let payload = PaginatedUpsertSignaturePayload {
-			payload,
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: block_number,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		System::set_block_number(block_number);
-		assert_err!(
-			StatefulStoragePallet::upsert_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::ProofHasExpired
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_signature_having_out_of_window_payload_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let payload = PaginatedUpsertSignaturePayload {
-			payload,
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: (<Test as Config>::MortalityWindowSize::get() + 1).into(),
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::upsert_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::ProofNotYetValid
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_signature_having_wrong_signature_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let (signature_key, _) = sr25519::Pair::generate();
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let payload = PaginatedUpsertSignaturePayload {
-			payload,
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = signature_key.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::upsert_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidSignature
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_signature_having_non_existing_msa_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let pair = get_invalid_msa_signature_account(); // hardcoded key that returns None Msa
-		let delegator_key = pair.public();
-		let msa_id = 1;
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let payload = PaginatedUpsertSignaturePayload {
-			payload,
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::upsert_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidMessageSourceAccount
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_signature_having_wrong_msa_in_payload_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (_correct_msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let wrong_msa_id = 3;
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let payload = PaginatedUpsertSignaturePayload {
-			payload,
-			target_hash: PageHash::default(),
-			msa_id: wrong_msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::upsert_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidMessageSourceAccount
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_signature_having_invalid_schema_id_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = INVALID_SCHEMA_ID;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let payload = PaginatedUpsertSignaturePayload {
-			payload,
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::upsert_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidSchemaId
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_signature_having_invalid_schema_location_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = ITEMIZED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let payload = PaginatedUpsertSignaturePayload {
-			payload,
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::upsert_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::SchemaPayloadLocationMismatch
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_signature_having_invalid_hash_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let payload = PaginatedUpsertSignaturePayload {
-			payload,
-			target_hash: PageHash::default() + 1, // any non default hash value
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::upsert_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::StalePageState
-		)
-	})
-}
-
-#[test]
-fn upsert_page_with_signature_having_valid_inputs_should_work() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		let page: PaginatedPage<Test> = payload.clone().into();
-		let payload = PaginatedUpsertSignaturePayload {
-			payload,
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_ok!(StatefulStoragePallet::upsert_page_with_signature(
-			RuntimeOrigin::signed(caller_1),
-			delegator_key.into(),
-			owner_signature,
-			payload
-		));
-
-		// assert
-		let keys = (schema_id, page_id);
-		let new_page: PaginatedPage<Test> = <StatefulChildTree>::try_read(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			PAGINATED_STORAGE_PREFIX,
-			&keys,
-		)
-		.expect("new page could not be read")
-		.expect("new page is empty");
-		assert_eq!(page.data, new_page.data, "new page contents incorrect");
-		assert_eq!(new_page.nonce, 1, "new page nonce incorrect");
-		System::assert_last_event(
-			StatefulEvent::PaginatedPageUpdated {
-				msa_id,
-				schema_id,
-				page_id,
-				prev_content_hash: PageHash::default(),
-				curr_content_hash: new_page.get_hash(),
-			}
-			.into(),
-		);
-	})
-}
-
-#[test]
-fn delete_page_with_signature_having_page_id_out_of_bounds_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = (<Test as Config>::MaxPaginatedPageId::get() + 1).into();
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::delete_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::PageIdExceedsMaxAllowed
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_signature_having_expired_payload_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let block_number = 10;
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		System::set_block_number(block_number);
-		assert_err!(
-			StatefulStoragePallet::delete_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::ProofHasExpired
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_signature_having_out_of_window_payload_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: (<Test as Config>::MortalityWindowSize::get() + 1).into(),
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::delete_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::ProofNotYetValid
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_signature_having_wrong_signature_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let (signature_key, _) = sr25519::Pair::generate();
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = signature_key.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::delete_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidSignature
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_signature_having_non_existing_msa_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let pair = get_invalid_msa_signature_account(); // hardcoded key that returns None Msa
-		let delegator_key = pair.public();
-		let msa_id = 1;
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::delete_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidMessageSourceAccount
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_signature_having_wrong_msa_in_payload_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (_correct_msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let wrong_msa_id = 3;
-		let schema_id = UNDELEGATED_PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: PageHash::default(),
-			msa_id: wrong_msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::delete_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidMessageSourceAccount
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_signature_having_invalid_schema_id_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = INVALID_SCHEMA_ID;
-		let page_id = 1;
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::delete_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::InvalidSchemaId
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_signature_having_invalid_schema_location_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = ITEMIZED_SCHEMA;
-		let page_id = 1;
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::delete_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::SchemaPayloadLocationMismatch
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_signature_having_invalid_hash_should_fail() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(Some(100));
-		assert_ok!(StatefulStoragePallet::upsert_page(
-			RuntimeOrigin::signed(caller_1.clone()),
-			msa_id,
-			schema_id,
-			page_id,
-			NONEXISTENT_PAGE_HASH,
-			payload.into(),
-		));
-
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_err!(
-			StatefulStoragePallet::delete_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Error::<Test>::StalePageState
-		)
-	})
-}
-
-#[test]
-fn delete_page_with_signature_with_non_existing_page_should_noop() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 1;
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: PageHash::default(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_storage_noop!(assert_eq!(
-			StatefulStoragePallet::delete_page_with_signature(
-				RuntimeOrigin::signed(caller_1),
-				delegator_key.into(),
-				owner_signature,
-				payload
-			),
-			Ok(())
-		));
-	})
-}
-
-#[test]
-fn delete_page_with_signature_having_valid_inputs_should_remove_page() {
-	new_test_ext().execute_with(|| {
-		// arrange
-		let caller_1 = test_public(1);
-		let (msa_id, pair) = get_signature_account();
-		let delegator_key = pair.public();
-		let schema_id = PAGINATED_SCHEMA;
-		let page_id = 1;
-		let page = generate_page::<PaginatedPageSize>(Some(1), Some(100));
-		<StatefulChildTree>::write(
-			&msa_id,
-			PALLET_STORAGE_PREFIX,
-			PAGINATED_STORAGE_PREFIX,
-			&(schema_id, page_id),
-			&page,
-		);
-
-		let payload = PaginatedDeleteSignaturePayload {
-			target_hash: page.get_hash(),
-			msa_id,
-			expiration: 10,
-			schema_id,
-			page_id,
-		};
-		let encode_data_new_key_data = wrap_binary_data(payload.encode());
-		let owner_signature: MultiSignature = pair.sign(&encode_data_new_key_data).into();
-
-		// act
-		assert_ok!(StatefulStoragePallet::delete_page_with_signature(
-			RuntimeOrigin::signed(caller_1),
-			delegator_key.into(),
-			owner_signature,
-			payload
-		));
-
-		// assert
-		let removed_page: Option<PaginatedPage<Test>> =
-			StatefulChildTree::<<Test as Config>::KeyHasher>::try_read(
-				&msa_id,
-				PALLET_STORAGE_PREFIX,
-				PAGINATED_STORAGE_PREFIX,
-				&(schema_id, page_id),
-			)
-			.unwrap();
-		assert!(removed_page.is_none());
-		System::assert_last_event(
-			StatefulEvent::PaginatedPageDeleted {
-				msa_id,
-				schema_id,
-				page_id,
-				prev_content_hash: page.get_hash(),
-			}
-			.into(),
-		);
-	})
-}
-
-#[test]
 fn apply_delete_item_on_append_only_fails() {
 	new_test_ext().execute_with(|| {
 		// arrange
@@ -2420,41 +324,6 @@ fn apply_delete_item_on_append_only_fails() {
 				schema_id,
 				content_hash,
 				BoundedVec::try_from(actions2).unwrap(),
-			),
-			Error::<Test>::UnsupportedOperationForSchema
-		);
-	});
-}
-
-#[test]
-fn delete_page_fails_for_append_only() {
-	new_test_ext().execute_with(|| {
-		// setup
-		let caller_1 = test_public(1);
-		let msa_id = 1;
-		let schema_id = PAGINATED_APPEND_ONLY_SCHEMA;
-		let page_id = 11;
-		let payload = generate_payload_bytes::<PaginatedPageSize>(None);
-		let page: PaginatedPage<Test> = payload.clone().into();
-		let page_hash = page.get_hash();
-
-		// assert
-		assert_ok!(StatefulStoragePallet::upsert_page(
-			RuntimeOrigin::signed(caller_1.clone()),
-			msa_id,
-			schema_id,
-			page_id,
-			NONEXISTENT_PAGE_HASH,
-			payload.into(),
-		));
-
-		assert_err!(
-			StatefulStoragePallet::delete_page(
-				RuntimeOrigin::signed(caller_1),
-				msa_id,
-				schema_id,
-				page_id,
-				page_hash
 			),
 			Error::<Test>::UnsupportedOperationForSchema
 		);
